ingest/requirements.txt
flask
functions-framework==3.*
google-cloud-bigquery
google-cloud-logging
google-cloud-pubsub
google-cloud-storage
pandas
pyyaml